requirements.txt
numpy
matplotlib
pymc >= 5.0.0
pytensor
arviz
numpyro
ipywidgets
graphviz
diffrax
optax
jaxopt
jax2pytensor @ git+https://github.com/jdehning/jax2pytensor@main

